btrecord output option
Added the option to not do the generated I/O ops, but instead output
their descriptions in the format expected by btreplay. This is supported
by what is essentially btrecord hacked into the form of a Python
extension module.

The new option is activated with "-p outfile". Currently outfile is
assumed to exist, and it's probably handy to make it a FIFO.