fix for running out of ops
If the last op in the source trace was of a class not previously seen,
it would not be a key in the matrix, and so we would throw a KeyError
after running that op. Handle this case by catching the exception and
setting the next op to the initial one.